+++ R/paged.R
@@ -28,7 +28,8 @@ html_paged = function(
     ..., css = css, theme = theme, template = template, .pagedjs = TRUE,
     .pandoc_args = c(
       lua_filters('uri-to-fn.lua', 'loft.lua', 'footnotes.lua'), # uri-to-fn.lua must come before footnotes.lua
-      if (!is.null(csl)) c('--csl', csl)
+      if (!is.null(csl)) c('--csl', csl),
+      pandoc_chapter_name_args()
     )
   )
 }
@@ -91,6 +92,20 @@ jss_paged = function(
   jss_format
 }
 
+#' Create a paged HTML thesis document suitable for printing
+#'
+#' This output format is similar to \code{\link{html_paged}}. The only
+#' difference is in the default stylesheets and Pandoc template. See
+#' \url{https://pagedown.rbind.io/thesis-paged/} for an example.
+#' @param ...,css,template Arguments passed to \code{\link{html_paged}()}.
+#' @return An R Markdown output format.
+#' @export
+thesis_paged = function(
+  ..., css = c('thesis'), template = pkg_resource('html', 'thesis.html')
+) {
+  html_paged(..., css = css, template = template)
+}
+
 pagedown_dependency = function(css = NULL, js = FALSE) {
   list(htmltools::htmlDependency(
     'paged', packageVersion('pagedown'), src = pkg_resource(),
@@ -115,4 +130,30 @@ html_format = function(
   html_document2(
     ..., css = css, template = template, pandoc_args = c(.pandoc_args, pandoc_args)
   )
+}
+
+chapter_name = function() {
+  config = bookdown:::load_config()
+  chapter_name = config[['chapter_name']] %n% bookdown:::ui_language('chapter_name')
+  if(is.null(chapter_name) || identical(chapter_name, '')) return(NULL)
+  if(is.character(chapter_name)) {
+    if(length(chapter_name) <= 2) {
+      return(chapter_name)
+    } else {
+      stop('chapter_name must be of length 1 or 2')
+    }
+  } else {
+    stop('chapter_name in _bookdown.yml must be a character string')
+  }
+}
+
+pandoc_metadata_arg = function(name, value) {
+  if(!missing(value) && is.character(value)) {
+    value = deparse(value)
+  }
+  c('--metadata', if (missing(value)) name else paste0(name, '=', value))
+}
+
+pandoc_chapter_name_args = function() {
+  unlist(lapply(chapter_name(), pandoc_metadata_arg, name = 'chapter_name'))
 }